13/01 v3 fixing problems

+++ src/moneycalculator/Model/Number.java
@@ -107,6 +107,6 @@ public class Number {
 
     @Override
     public String toString() {
-        return String.valueOf(numerator / denominator);
+        return String.valueOf((double)numerator/(double)denominator);
     }
 }
